[Add] icons for programmig language, unknown files, folders

+++ layouts/files.php
@@ -1,20 +1,26 @@
 <?php
 
-function defineIcon($handle) {
+function defineIcon($handle, $path) {
     $ext = substr($handle, strrpos($handle, '.') + 1);
     $icon = "http://localhost/H5AI/icons/" . $ext . ".png" ;
-    echo "<p><img src=\"$icon\" alt=\"language icon\"/> ";    
+    if (file_exists("./icons/" . $ext . ".png"))
+        {
+            echo "<p><img src=\"$icon\" alt=\"language icon\"/> ";
+        }
+    else if (is_dir($path . '/' . $handle))
+        echo "<p><img src=\"http://localhost/H5AI/icons/dir.png\" alt=\"language icon\"/>" ;    
+    else
+        echo "<p><img src=\"http://localhost/H5AI/icons/unknown.png\" alt=\"language icon\"/>" ;
 }
 
-function displayDirContent($folders) {
+function displayDirContent($folders, $path) {
     $baseURL = "http://localhost" . $_SERVER['REQUEST_URI']; 
     while ($handle = readdir($folders)) {
         $url = $baseURL . "/" . $handle;
         if ($handle != "." && $handle != "..")
         {
-            if (strpos($handle, "."))
-                defineIcon($handle);
-            if (is_dir($handle))
+            defineIcon($handle, $path);
+            if (is_dir($path . "/" . $handle))
                 echo "<a href=\"{$url}\">{$handle}</a></p>";
             else
             echo "{$handle}</p>";
